review: rename dir from prover to tx_prover
